unique colour count for checking near misses in the game
