+++ mastermind.rb
@@ -1,0 +1,4 @@
+# mastermind.rb
+
+class Mastermind
+end
